Refactor import statements in token-info.tsx and load-step.tsx
