+++ views/new-migrate/migration-steps/step-1/token-info.tsx
@@ -1,4 +1,4 @@
-import SMLoader from "@/components/loader";
+import { SMLoader } from "@/components";
 
 const TokenInfo = ({
   decimal,
